requirements.txt
aiomysql==0.2.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
bcrypt==3.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.6.12
cryptography==43.0.3
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.8
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
mysqlclient==2.2.6
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
python-telegram-bot==21.9
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.1
